requirements.txt
Flask==3.0.2
rembg==2.0.55
Pillow==10.2.0
Flask-Cors==4.0.0
Werkzeug==3.0.1
